requirements.txt
google-cloud-logging
google-cloud-monitoring
pandas
pyarrow
openpyxl
